+++ Runtime/Adapters/FlexalonColliderAdapter.cs
@@ -77,7 +77,7 @@ namespace Flexalon
             }
 
             scale = Vector3.one;
-            return false;
+            return true;
         }
 
         public bool TryGetRectSize(FlexalonNode node, out Vector2 rectSize)
@@ -88,12 +88,12 @@ namespace Flexalon
 
         protected override void UpdateProperties()
         {
-            _node.SetAdapter(this);
+            Node.SetAdapter(this);
         }
 
         protected override void ResetProperties()
         {
-            _node.SetAdapter(null);
+            Node.SetAdapter(null);
         }
     }
 }